move new class into own file
